fix(abi): correct system-flag fn param widths to match C
Extend the flag-family param-width reconciliation. Per the C prototypes,
setSystemFlag/clearSystemFlag take `unsigned int` and fnSetFlag takes `uint16_t`;
several owners disagreed:

- clearSystemFlag(flag: u16) -> c_uint in frontier_integers (was narrower than
  the C unsigned int, same latent >0xFFFF truncation as setSystemFlag).
- math_command_wrappers_runtime bound the trio as i32: setSystemFlag(i32) and
  clearSystemFlag(i32) -> c_uint (signed 4-byte for an unsigned-int param), and
  fnSetFlag(i32) -> u16 (a 4-byte binding for the 2-byte uint16_t param).

The seven fnSetFlag(u16) owners already match C uint16_t and are left as-is. Call
sites pass flag constants that coerce cleanly to the corrected types.

Verified: zig build sim and zig build test exit 0 (9674 testSuite),
math_command_wrappers_parity green, fmt clean.

+++ zig_src/frontier/frontier_integers.zig
@@ -80,7 +80,7 @@ extern var lastIntegerBase: u32;
 extern var temporaryInformation: u8;
 
 extern fn setSystemFlag(flag: c_uint) void;
-extern fn clearSystemFlag(flag: u16) void;
+extern fn clearSystemFlag(flag: c_uint) void;
 
 const c_moreInfoOnError = @extern(*const fn (m1: [*:0]const u8, m2: ?[*:0]const u8, m3: ?[*:0]const u8, m4: ?[*:0]const u8) callconv(.c) void, .{ .name = "moreInfoOnError" });
